% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoneseasonwiki.R
\name{change_colnames}
\alias{change_colnames}
\title{Short function for changing column names after downloading the data.}
\usage{
change_colnames(df)
}
\arguments{
\item{df}{Data frame returned by get_one_season_wiki function after manipulations from datadownload.R file.}
}
\value{
Data frame df with changed colnames.
}
\description{
Short function for changing column names after downloading the data.
}

