% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimdboverothers.R
\name{change_wday}
\alias{change_wday}
\title{Change day of the week to given day of the week.}
\usage{
change_wday(actual_date, target_wday)
}
\arguments{
\item{actual_date}{Date to be modified by changing day of the week.}

\item{target_wday}{Day of the week as coded in lubridate package.}
}
\value{
Date given date with day chosen to closest given day of the week.
}
\description{
Change day of the week to given day of the week.
}

