% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoneseasonwiki.R
\name{changeColnames}
\alias{changeColnames}
\title{Short function for changing column names after downloading the data.}
\usage{
changeColnames(df)
}
\arguments{
\item{df}{Data frame returned by getOneSeasonWiki function after manipulations from datadownload.R file.}
}
\value{
Data frame df with changed colnames.
}
\description{
Short function for changing column names after downloading the data.
}

