% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotratings.R
\name{changeWday}
\alias{changeWday}
\title{Helper function to change day of the week to given day of the week.}
\usage{
changeWday(actualDate, targetWday)
}
\arguments{
\item{actualDate}{Date to be modified by changing day of the week.}

\item{targetWday}{Day of the week as coded in lubridate package.}
}
\value{
Date given date with day chosen to closest given day of the week.
}
\description{
Helper function to change day of the week to given day of the week.
}

