% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnetflix.R
\name{filterNetflix}
\alias{filterNetflix}
\title{Filter data to plot Netflix show's ratings or compare it to other show.}
\usage{
filterNetflix(showNames, seasons, minRating = 0, maxRating = 10)
}
\arguments{
\item{showNames}{chr, full titles of shows to plot.}

\item{seasons}{chr, one season for each show in showNames.}

\item{minRating}{dbl, minimum rating to plot.}

\item{maxRating}{db, maximum rating to plot.}
}
\value{
Data frame with added column with episodes signature seasonxepisode.
}
\description{
Filter data to plot Netflix show's ratings or compare it to other show.
}

