% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotoneshow.R
\name{filterOneShow}
\alias{filterOneShow}
\title{Filter data to plot given type of ratings (Nielsen/IMDb) for a given show.}
\usage{
filterOneShow(showName, typeRating, seasons = NULL, minRating = 0,
  maxRating = 10, minDate = defaultMinDate, maxDate = defaultMaxDate)
}
\arguments{
\item{showName}{Character with show's full name.}

\item{typeRating}{Character with type of ratings
that matches one of column names in episodes tibble.}

\item{seasons}{Integer vector with numbers of seasons to display.}

\item{minRating}{Double minimum ratings value to display.}

\item{maxRating}{Double miximum ratings value to display.}

\item{minDate}{Date earliest date to display.}

\item{maxDate}{Date most recent date to display.}
}
\value{
Tibble which is filtered episodes tibble with added
        'key' column for plotting function.
}
\description{
Filter data to plot given type of ratings (Nielsen/IMDb) for a given show.
}

