% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotratings.R
\name{filterToPlot}
\alias{filterToPlot}
\title{Filter data for plotting function.}
\usage{
filterToPlot(showNames, typeRating, seasons = 1:defaultSeasons,
  minRating = 0, maxRating = 10, minDate = defaultMinDate,
  maxDate = defaultMaxDate)
}
\arguments{
\item{showNames}{chr vector with one or two full names of shows.}

\item{typeRating}{chr, one of vales "imdbRating", "nielsenRating".}

\item{seasons}{int vector with numbers of season to display.}

\item{minRating}{dbl, only episodes with higher rating will be displayed.}

\item{maxRating}{dbl, only episodes with lower rating will be displayed.}

\item{minDate}{date, minimum air date to display.}

\item{maxDate}{date, maximum air date to display.}
}
\value{
list of two data frame: first with data for background line,
        second with data for chosen shows.
}
\description{
Filter data for plotting function.
}

