% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodesinfo.R
\name{get_episodes_info}
\alias{get_episodes_info}
\title{Information about given episodes extracted from IMDb pages.}
\usage{
get_episodes_info(eps_links)
}
\arguments{
\item{eps_link}{Links to episodes returned from get_episodes_links function.}
}
\value{
Data frame with columns show (character, name of the show), cancelled (logical, TRUE if show has been cancelled),
        season (character, season number), episode (episode numebr within season),
        air_date (date, air date), rating (numeric, IMDb rating), num_of_votes (integer, number of rating votes).
}
\description{
Information about given episodes extracted from IMDb pages.
}

