% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodelinks.R
\name{get_episodes_links}
\alias{get_episodes_links}
\title{Links to all episodes of one show.}
\usage{
get_episodes_links(seasons)
}
\arguments{
\item{seasons}{Character vector of links to season episodes as returned by get_seasons_links function.}
}
\value{
List. Each element of the list is a vector with links to episodes of one season.
}
\description{
Links to all episodes of one show.
}

