% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getratingsrt.R
\name{get_ratings_rt}
\alias{get_ratings_rt}
\title{Download ratings from one RottenTomatoes page.}
\usage{
get_ratings_rt(link)
}
\arguments{
\item{link}{Link to RT page.}
}
\description{
Download ratings from one RottenTomatoes page.
}

