% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinks.R
\name{get_seasons_links}
\alias{get_seasons_links}
\title{Get links to all seasons for a given show.}
\usage{
get_seasons_links(show_link)
}
\arguments{
\item{show_link}{Link to the show's main page in IMDb.}
}
\value{
Vector of full links to show's seasons.
}
\description{
Get links to all seasons for a given show.
}

