% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodesinfo.R
\name{getEpisodesInfo}
\alias{getEpisodesInfo}
\title{Information about given episodes extracted from IMDb pages.}
\usage{
getEpisodesInfo(epLinks)
}
\arguments{
\item{epLinks}{Links to episodes returned from get_episodes_links function.}
}
\value{
Data frame with columns show (character, name of the show), cancelled (logical, TRUE if show has been cancelled),
        season (character, season number), episode (episode numebr within season),
        air_date (date, air date), rating (numeric, IMDb rating), num_of_votes (integer, number of rating votes).
}
\description{
Information about given episodes extracted from IMDb pages.
}

