% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodelinks.R
\name{getEpisodesLinks}
\alias{getEpisodesLinks}
\title{Links to all episodes of one show.}
\usage{
getEpisodesLinks(seasons)
}
\arguments{
\item{seasons}{Character vector of links to season episodes as returned by getSeasonLinks function.}
}
\value{
List. Each element of the list is a vector with links to episodes of one season.
}
\description{
Links to all episodes of one show.
}

