% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoneseasonwiki.R
\name{getOneSeasonWiki}
\alias{getOneSeasonWiki}
\title{Download ratings table from wikipedia page.}
\usage{
getOneSeasonWiki(link, selector)
}
\arguments{
\item{link}{Url of a wikipedia page that containts the ratings table.}

\item{selector}{Unique css selector of ratings table.}
}
\value{
List returned by rvest html_table function.
}
\description{
Download ratings table from wikipedia page.
}

