% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getratingsrt.R
\name{getRatingsRT}
\alias{getRatingsRT}
\title{Download ratings from one RottenTomatoes page.}
\usage{
getRatingsRT(link)
}
\arguments{
\item{link}{Link to RT page.}
}
\description{
Download ratings from one RottenTomatoes page.
}

