% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinks.R
\name{getSeasonsLinks}
\alias{getSeasonsLinks}
\title{Get links to all seasons for a given show.}
\usage{
getSeasonsLinks(showLink, noOfSeasons)
}
\arguments{
\item{showLink}{Link to the show's main page in IMDb.}

\item{noOfSeasons}{Number of show's seasons.}
}
\value{
Vector of full links to show's seasons.
}
\description{
Get links to all seasons for a given show.
}

