% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givedetails.R
\name{giveShortDetails}
\alias{giveShortDetails}
\title{Display shorter episode info under comparison plot.}
\usage{
giveShortDetails(filtData, click, seasons)
}
\arguments{
\item{click}{Shiny on_click object.}

\item{seasons}{numeric, vector of plotted seasons numbers.}

\item{filData}{Data frame returned by filterToPlot or filterNetflix function.}
}
\value{
character with show title, episode signature, title and IMDb ratings info.
}
\description{
Display shorter episode info under comparison plot.
}

