% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givetooltip.R
\name{giveTooltip}
\alias{giveTooltip}
\title{Text info about given episode.}
\usage{
giveTooltip(filtData, click, typeRating)
}
\arguments{
\item{filtData}{Data frame returned by filterToPlot function.}

\item{click}{shiny object returned on plot click.}

\item{typeRating}{chr, "imdbRating" or "nielsenRating".}
}
\value{
html with show name, episode's signature, episode's title,
        rating and number of votes/numbers of viewers, air date in 
        separate lines.
}
\description{
Text info about given episode.
}

