% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodesinfo.R
\name{one_episode_info}
\alias{one_episode_info}
\title{Information about one episode.}
\usage{
one_episode_info(ep_link)
}
\arguments{
\item{ep_link}{Character string: link to episode's IMDb page.}
}
\value{
vector containing show name, value indicating if the show was cancelled,
               season and episode number, air date, episode title, rating and number of votes.
}
\description{
Information about one episode.
}

