% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getepisodesinfo.R
\name{oneEpisodeInfo}
\alias{oneEpisodeInfo}
\title{Information about one episode.}
\usage{
oneEpisodeInfo(epLink)
}
\arguments{
\item{epLink}{Character string: link to episode's IMDb page.}
}
\value{
vector containing show name, value indicating if the show was cancelled,
               season and episode number, air date, episode title, rating and number of votes.
}
\description{
Information about one episode.
}

