% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimdboverothers.R
\name{plot_imdb_over_others}
\alias{plot_imdb_over_others}
\title{Plot IMDb ratings of one show with other show's mean rating in the background.}
\usage{
plot_imdb_over_others(show_name)
}
\arguments{
\item{show_name}{Name of the show to plot.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings of one show with other show's mean rating in the background.
}

