% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimdboverothers.R
\name{plot_imdb_ratings}
\alias{plot_imdb_ratings}
\title{Plot IMDb ratings of one show with other show's mean rating in the background.}
\usage{
plot_imdb_ratings(show_name, background = FALSE)
}
\arguments{
\item{show_name}{Name of the show to plot.}

\item{Logical.}{If TRUE, ratings for other shows will be displayed.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings of one show with other show's mean rating in the background.
}

