% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimdboverothers.R
\name{plotImdbRatings}
\alias{plotImdbRatings}
\title{Plot IMDb ratings of one show with other show's mean rating in the background.}
\usage{
plotImdbRatings(showName, background = FALSE)
}
\arguments{
\item{showName}{Name of the show to plot.}

\item{background}{logical, if TRUE, ratings for other shows will be displayed.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings of one show with other show's mean rating in the background.
}

