% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotoneshow.R
\name{plotOneShow}
\alias{plotOneShow}
\title{Plot one show's rating in time.}
\usage{
plotOneShow(sourceT)
}
\arguments{
\item{sourceT}{Tibble returned by filterOneShow function.}
}
\value{
ggvis plot
}
\description{
Plot one show's rating in time.
}

