% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotratings.R
\name{plotRatingsCompareVS}
\alias{plotRatingsCompareVS}
\title{Plot IMDb ratings vs Nielsen ratings for two shows.}
\usage{
plotRatingsCompareVS(sources, trend = FALSE)
}
\arguments{
\item{sources}{Data frame returned by filterToPlot function.}

\item{trend}{lgl, if TRUE, trend line will be displayed.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings vs Nielsen ratings for two shows.
}

