% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinksrt.R
\name{seasons_links_paste}
\alias{seasons_links_paste}
\title{Function for creating links to show's seasons on RottenTomatoes.}
\usage{
seasons_links_paste(shows_info_df)
}
\arguments{
\item{shows_info_df}{Data frame created in datadownload.R file.}
}
\description{
Function for creating links to show's seasons on RottenTomatoes.
}

