% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinksrt.R
\name{seasons_links_rt}
\alias{seasons_links_rt}
\title{Function for creating data frame with links to all shows' seasons from RT.}
\usage{
seasons_links_rt(shows_info_df)
}
\arguments{
\item{shows_info_df}{Data frame created in datadownload.R file.}
}
\description{
Function for creating data frame with links to all shows' seasons from RT.
}

