% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinksrt.R
\name{seasonsLinksPaste}
\alias{seasonsLinksPaste}
\title{Function for creating links to show's seasons on RottenTomatoes.}
\usage{
seasonsLinksPaste(showsInfoDf)
}
\arguments{
\item{showsInfoDf}{Data frame created in datadownload.R file.}
}
\description{
Function for creating links to show's seasons on RottenTomatoes.
}

