% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseasonslinksrt.R
\name{seasonsLinksRT}
\alias{seasonsLinksRT}
\title{Function for creating data frame with links to all shows' seasons from RT.}
\usage{
seasonsLinksRT(showsInfoDf)
}
\arguments{
\item{showsInfoDf}{Data frame created in datadownload.R file.}
}
\description{
Function for creating data frame with links to all shows' seasons from RT.
}

