% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shows}
\alias{shows}
\title{Data on shows' seasons.}
\format{Data frame with 23 rows and 9 columns,
  \describe{
    \item{show}{chr, full show name (excluding "Marvel's" or "DC's")}
    \item{season}{chr, season number}
    \item{episode}{chr, episode number}
    \item{cancelled}{lgl, TRUE if the show is now cancelled}
    \item{channel}{chr, name of channel on which the show air or Netflix for Netflix shows}
    \item{comic}{chr, name of comic book publisher on whose comic books the show is based}
    \item{critics}{int, percentage of "fresh" ratings from critics on RottenTomatoes}
    \item{audience}{int, percentage of "fresh" ratings from audience on RT}
    \item{criticsAveraged}{dbl, average score from critics on scale 1-10 from RT}
    \item{audienceAverage}{dbl, average score from audience on scale 1-5 from RT}
  }
 @source \url{https://imdb.com}, \url{https://wikipedia.org},
         \url{https://rottentomatoes.com}, \url{https://tvseriesfinale.com}}
\usage{
shows
}
\description{
Data on shows' seasons.
}
\keyword{datasets}

