% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givetooltip.R
\name{giveTooltip}
\alias{giveTooltip}
\title{Function that returns a tooltip for plotOneShow plot.}
\usage{
giveTooltip(key)
}
\arguments{
\item{key}{Identifier from plotOneShow function.}
}
\value{
HTML generating tooltip.
}
\description{
Function that returns a tooltip for plotOneShow plot.
}

