% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnetflix.R
\name{plotNetflix}
\alias{plotNetflix}
\title{Plot Netflix show's rating or compare it with other show.}
\usage{
plotNetflix(sourceT, trend = FALSE)
}
\arguments{
\item{sourceT}{Data frame returned by filterNetflix function.}

\item{trend}{lgl, if TRUE, trend line will be ploted.}
}
\value{
ggplot2 object.
}
\description{
Plot Netflix show's rating or compare it with other show.
}

