% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimdboverothers.R
\name{plotRatings}
\alias{plotRatings}
\title{Plot IMDb ratings of one show.}
\usage{
plotRatings(background = TRUE, trend = FALSE, ...)
}
\arguments{
\item{background}{logical, if TRUE, ratings for other shows will be displayed.}

\item{trend}{logical, if TRUE, trend line for season is plotted.}

\item{...}{arguments passed to filterToPlot function.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings of one show.
}

