% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotratings.R
\name{plotRatings}
\alias{plotRatings}
\title{Plot IMDb ratings of one show.}
\usage{
plotRatings(sources, background = FALSE, trend = FALSE, separate = FALSE)
}
\arguments{
\item{sources}{list returned by filterToPlot function.}

\item{background}{logical, if TRUE, ratings for other shows will be displayed.}

\item{trend}{logical, if TRUE, trend line for season is plotted.}

\item{separated}{lgl, if TRUE, facets will be used.}
}
\value{
ggplot2 object.
}
\description{
Plot IMDb ratings of one show.
}

